move instrumentref parameter to specialized module
