+++ qcodes/parameters/parameter.py
@@ -33,7 +33,7 @@ from qcodes.data.data_array import DataArray
 from qcodes.utils.command import Command
 from qcodes.utils.helpers import full_class, is_sequence_of, warn_units
 from qcodes.utils.metadata import Metadatable
-from qcodes.utils.validators import Strings, Validator
+from qcodes.utils.validators import Validator
 
 from .parameter_base import ParamDataType, ParameterBase, ParamRawDataType
 from .sweep_values import SweepFixedValues
@@ -1301,76 +1301,6 @@ class CombinedParameter(Metadatable):
             meta_data[str(parameter)] = parameter.snapshot()
 
         return meta_data
-
-
-class InstrumentRefParameter(Parameter):
-    """
-    An instrument reference parameter.
-
-    This parameter is useful when one needs a reference to another instrument
-    from within an instrument, e.g., when creating a meta instrument that
-    sets parameters on instruments it contains.
-
-    Args:
-        name: The name of the parameter that one wants to add.
-
-        instrument: The "parent" instrument this
-            parameter is attached to, if any.
-
-        initial_value: Starting value, may be None even if None does not
-            pass the validator. None is only allowed as an initial value
-            and cannot be set after initiation.
-
-        **kwargs: Passed to InstrumentRefParameter parent class
-    """
-
-    def __init__(
-        self,
-        name: str,
-        instrument: Optional["InstrumentBase"] = None,
-        label: Optional[str] = None,
-        unit: Optional[str] = None,
-        get_cmd: Optional[Union[str, Callable[..., Any], Literal[False]]] = None,
-        set_cmd: Optional[Union[str, Callable[..., Any], Literal[False]]] = None,
-        initial_value: Optional[Union[float, str]] = None,
-        max_val_age: Optional[float] = None,
-        vals: Optional[Validator[Any]] = None,
-        docstring: Optional[str] = None,
-        **kwargs: Any,
-    ) -> None:
-        if vals is None:
-            vals = Strings()
-        if set_cmd is not None:
-            raise RuntimeError("InstrumentRefParameter does not support " "set_cmd.")
-        super().__init__(
-            name,
-            instrument,
-            label,
-            unit,
-            get_cmd,
-            set_cmd,
-            initial_value,
-            max_val_age,
-            vals,
-            docstring,
-            **kwargs,
-        )
-
-    # TODO(nulinspiratie) check class works now it's subclassed from Parameter
-    def get_instr(self) -> "InstrumentBase":
-        """
-        Returns the instance of the instrument with the name equal to the
-        value of this parameter.
-        """
-        ref_instrument_name = self.get()
-        # note that _instrument refers to the instrument this parameter belongs
-        # to, while the ref_instrument_name is the instrument that is the value
-        # of this parameter.
-        if self._instrument is None:
-            raise RuntimeError(
-                "InstrumentRefParameter is not bound to " "an instrument."
-            )
-        return self._instrument.find_instrument(ref_instrument_name)
 
 
 class ManualParameter(Parameter):
